# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Body_create_document_api_v1_management_knowledge_documents_post": {
        "properties": {
          "categories": {
            "title": "Categories",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "review_after": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review After"
          },
          "slug": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Slug",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/KnowledgeSourceType"
          },
          "source_urls": {
            "default": "[]",
            "title": "Source Urls",
            "type": "string"
          },
          "title": {
            "maxLength": 240,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          },
          "version": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "file",
          "slug",
          "title",
          "version",
          "source_type",
          "categories",
          "verified_at"
        ],
        "title": "Body_create_document_api_v1_management_knowledge_documents_post",
        "type": "object"
      },
      "Body_create_document_version_api_v1_management_knowledge_documents__document_id__versions_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "review_after": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review After"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          },
          "version": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "file",
          "version",
          "verified_at"
        ],
        "title": "Body_create_document_version_api_v1_management_knowledge_documents__document_id__versions_post",
        "type": "object"
      },
      "Category": {
        "enum": [
          "BLOQUEO_TARJETA",
          "REPORTE_FRAUDE",
          "CONSULTA_GENERAL",
          "SOLICITUD_CREDITO",
          "BANCA_DIGITAL"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeDocumentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeDocumentSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "KnowledgeDocumentPage",
        "type": "object"
      },
      "KnowledgeDocumentSummary": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "title": "Categories",
            "type": "array"
          },
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "index_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Error"
          },
          "index_status": {
            "$ref": "#/components/schemas/KnowledgeIndexStatus"
          },
          "indexed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "review_after": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review After"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/KnowledgeSourceType"
          },
          "source_urls": {
            "items": {
              "type": "string"
            },
            "title": "Source Urls",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "version",
          "source_type",
          "categories",
          "source_urls",
          "verified_at",
          "review_after",
          "file_name",
          "mime_type",
          "byte_size",
          "page_count",
          "content_sha256",
          "index_status",
          "indexed_at",
          "index_error",
          "active",
          "chunk_count",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeDocumentSummary",
        "type": "object"
      },
...
      "KnowledgeIndexStatus": {
        "enum": [
          "PENDING",
          "INDEXING",
          "READY",
          "FAILED",
          "ARCHIVED"
        ],
        "title": "KnowledgeIndexStatus",
        "type": "string"
      },
...
      "KnowledgeJobResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/KnowledgeDocumentSummary"
          },
          "job": {
            "$ref": "#/components/schemas/KnowledgeJobSummary"
          }
        },
        "required": [
          "job",
          "document"
        ],
        "title": "KnowledgeJobResponse",
        "type": "object"
      },
...
      "KnowledgeSourceType": {
        "enum": [
          "OFFICIAL",
          "REGULATORY",
          "INTERNAL",
          "HYBRID"
        ],
        "title": "KnowledgeSourceType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema de orquestacion de atencion bancaria",
    "title": "Sistema de Orquestacion Bancaria",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/management/knowledge/documents": {
      "get": {
        "operationId": "list_documents_api_v1_management_knowledge_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "index_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeIndexStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Index Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "Conocimiento gerencial"
        ]
      },
      "post": {
        "operationId": "create_document_api_v1_management_knowledge_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_api_v1_management_knowledge_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Document",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
    "/api/v1/management/knowledge/documents/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_management_knowledge_documents_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
    "/api/v1/management/knowledge/documents/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_v1_management_knowledge_documents_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
    "/api/v1/management/knowledge/documents/{document_id}": {
      "delete": {
        "operationId": "archive_document_api_v1_management_knowledge_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Document",
        "tags": [
          "Conocimiento gerencial"
        ]
      },
      "get": {
        "operationId": "get_document_api_v1_management_knowledge_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "Conocimiento gerencial"
        ]
      },
      "patch": {
        "operationId": "update_document_api_v1_management_knowledge_documents__document_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Document",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
    "/api/v1/management/knowledge/documents/{document_id}/download": {
      "get": {
        "operationId": "download_document_api_v1_management_knowledge_documents__document_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Document",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
    "/api/v1/management/knowledge/documents/{document_id}/reindex": {
      "post": {
        "operationId": "reindex_document_api_v1_management_knowledge_documents__document_id__reindex_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reindex Document",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
    "/api/v1/management/knowledge/documents/{document_id}/versions": {
      "post": {
        "operationId": "create_document_version_api_v1_management_knowledge_documents__document_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_version_api_v1_management_knowledge_documents__document_id__versions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Document Version",
        "tags": [
          "Conocimiento gerencial"
        ]
      }
    },
...
  }
}